Add support for SLURM sacct command to retrieve job completion information.
The "scontrol show job" command is preferred and will always be used first.  
However, SLURM, by default, purges the records available with that command 
after 300 seconds.  If the job information can not found using "scontrol 
show job" Rocoto will use "sacct" to retrieve it.  Unfortunately, the use 
of sacct requires that SLURM is configured with AccountingStorage enabled.
If AccountingStorage is not enabled, there is no way to retrieve the job
records if more than 300 seconds (by default) has passed between when the job 
finished and when rocotorun is executed.  This means that SLURM can only
be fully supported for installations that have AccountingStorage enabled.