cycleexistdep branch: several new capabilities, plus WCOSS support. WCOSS is the National Weather Service pair of identical twin supercomputers for operational weather forecasting.  These changes have been tested on Jet, Zeus and WCOSS, with no trouble found so far.
1. Multiple <native> tags are allowed per task.  Within Rocoto, they
are no longer accessible as an attribute, instead you have to loop
over them via task.each_native.  This requires changes to all batch
system classes.  I think I have made those changes, but they need to
be tested.  However, any new batch systems, that are not in this
branch, need similar changes.  The loadlevelerbatchsystem.rb does not
appear to have support for native tags yet, so I did not modify it.

2. When using LSF, the lsfbatchsystem.rb will check the bjobs -l -a
output for completed jobs before running bhist.  This is needed on
WCOSS because jobs take a long time to appear in bhist.  This required
only minor changes because the output of bjobs -l and bhist are
similar.  With these changes, Rocoto will rarely have to resort to
bhist, so runtime should be shorter.

3. The <nodes> support was broken in LSF: when provided with X:ppn=Y,
the code pretended X=1.  That is now fixed.

4. The <cores> support was broken when a shared queue was in use.  In
that situation, the WCOSS LSF requires span[ptile=].  This update adds
an optional <nodesize> tag that, if present, will be used to calculate
the span[ptile=] argument.  If absent, the old behavior is used, which
will break on WCOSS.

5. The WCOSS esub job preparser breaks your job if
LSB_PJL_TASK_GEOMETRY is set.  This only happens to some jobs, with no
obvious pattern or fix.  IBM's suggested fix is to just not set that
variable before running bsub.  Instead, Rocoto now sets
ROCOTO_TASK_GEO, and the lsfwrapper.sh copies that value to
LSB_PJL_TASK_GEOMETRY after the job starts.

6. Extra junk puts debugging lines, and enabling STDOUT writing from
the server.  This needs to be removed before merging to the trunk.