Add the capability to specify an activation offset for cycles when running in realtime mode.  An optional attribute called "activation_offset" was added to the <cycledef> tag.  This attribute can be any time field, positive or negative, in the DD:HH:MM:SS format.  For example,
<cycledef group="test" activation_offset="-01:00:00">...</cycledef>

In retro mode, this attribute is ignored.  In realtime mode, this
causes the cycles to be activated when the system clock reaches
the cycle time + the activation offset.  A negative offset will cause
cycles to be activated before the system clock reaches the cycle time.
A positive offset will cause the cycles to be activated after the
system clock reaches the cycle time.  For example,

<cycledef activation_offset="01:00:00">...</cycledef>

will cause the 12Z cycle to be activated at 13Z.

And,

<cycledef activation_offset="-01:00:00">...</cycledef>

will cause the 12Z cycle to be activated at 11Z.

The purpose of this attribute is to allow for situations where a
realtime workflow's tasks must start before the system clock reaches
the cycle time.  It can also be used when it is known that a workflow's
cycles won't be able to start until some time after the system clock
reaches the cycle time due to input data delivery delays, for example.

The default value of activation_offset is 0.