Bug Fix:
ENV is not thread safe.  LSF job submission relied on the use of ENV
to pass environment variables to jobs.  However, since job submission
is multi-threaded, the modifications to ENV by the threads polluted
each others' environments.

This was fixed by wrapping the submission command in /bin/env with
a list of environment settings.  This allows the threads to work
concurrently without harming each other.