Allow Rocoto to work on read-only databases.  This will disable all functionality that requires writing, and will prevent Rocoto from creating tables or using the lock databases.  If the user tries to run any Rocoto command that needs write access, the command will complain and exit.
There is a warning message printed when the database is opened in
read-only mode.

Read-only detection happens at the top of the program by examining
the db and lock db files.  If files do not exist, they they're
assumed to be writable.  If they exist and are not writable by the
current user, then Rocoto drops to read-only mode.

Read-only can be queried by db.readonly?