Added locking of ~/.rocoto/log messaging to prevent race conditions that can cause the log to be removed or corrupted.
Added a WORKFLOW_ID to ~/.rocoto/log and stderr logging to help identify
which workflow is generating the error.  The WORKFLOW_ID is equal to the
name of the XML file in use (which might not be unique since full path is
not used).

Added retry capability for database lock and unlock attempts when those
operations fail due to SQLite exclusive locks.