Add TaskValidDependency
The TaskValidDependency is used to check whether a task is valid for

given cycle. This comes in handy when setting up complex
ndencies
for workflows in which the completion of a cycle depends on a task

may or may not have needed to run in that cycle.

An example:
You declare two cycledefs, prep and fcst, in which fcst runs a
deterministic forecast once daily, and requires a background ensemble to
run 6 hours prior to the forecast. The prep cycles run a
task named 'bkg_ens_stats', while the the fcst cycles run a task
'forecast'. Successful completion of a cycle depends (among other
things) on both of these tasks having completed, but not for every
cycle. A dependency then, for any valid cycle to be complete could
be:

<and>
  <or>
   <not><taskvalid task="bkg_ens_stats"/></not>
   <taskdep task="bkg_ens_stats"/>
  </or>
  <or>
   <not><taskvalid task="forecast"/></not>
   <taskdep task="forecast"/>
  </or>
</and>

This dependency would check whether a task should run and if it should,
that it completed successfully.